+++ src/main/java/net/grid/vampiresdelight/client/extension/PourableBottleItemExtension.java
@@ -7,8 +7,9 @@ import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.client.extensions.common.IClientItemExtensions;
+import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
 
+// TODO: Fix this and find how to use it with bottles as it's now off
 public class PourableBottleItemExtension  implements IClientItemExtensions {
     private static final HumanoidModel.ArmPose POSE = HumanoidModel.ArmPose.create(VampiresDelight.MODID + "_pouring_pose", true, (model, entity, arm) -> {
         boolean isRightArmPouring = arm == HumanoidArm.RIGHT;
@@ -29,11 +30,4 @@ public class PourableBottleItemExtension  implements IClientItemExtensions {
     public HumanoidModel.ArmPose getArmPose(LivingEntity entityLiving, InteractionHand hand, ItemStack itemStack) {
         return (!itemStack.isEmpty() && entityLiving.getUsedItemHand() == hand && entityLiving.getUseItemRemainingTicks() > 0) ? POSE : HumanoidModel.ArmPose.EMPTY;
     }
-
-    /*
-    @Override
-    public BlockEntityWithoutLevelRenderer getCustomRenderer() {
-        return new PourableBottleItemRenderer();
-    }
-     */
 }