+++ src/main/java/net/grid/vampiresdelight/common/event/MappingEventHandler.java
@@ -2,24 +2,32 @@ package net.grid.vampiresdelight.common.event;
 
 import net.grid.vampiresdelight.VampiresDelight;
 import net.grid.vampiresdelight.common.registry.VDItems;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.MissingMappingsEvent;
-import net.minecraft.world.item.Item;
+import net.minecraft.resources.ResourceLocation;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NewRegistryEvent;
 import org.jetbrains.annotations.NotNull;
 
-@Mod.EventBusSubscriber(modid = VampiresDelight.MODID)
+@EventBusSubscriber(modid = VampiresDelight.MODID)
 public class MappingEventHandler {
     @SubscribeEvent
-    public static void onMissingMappings(@NotNull MissingMappingsEvent event) {
-        event.getAllMappings(ForgeRegistries.Keys.ITEMS).forEach(MappingEventHandler::fixItems);
+    public static void fix(NewRegistryEvent event) {
+        fixItems(new Mapping(VDItems.ITEMS));
     }
 
-    public static void fixItems(@NotNull MissingMappingsEvent.Mapping<Item> mapping) {
-        switch (mapping.getKey().toString()) {
-            case "vampiresdelight:eye_toast" -> mapping.remap(VDItems.EYE_CROISSANT.get());
-            case "vampiresdelight:wine_glass" -> mapping.remap(VDItems.BLOOD_WINE_GLASS.get());
+    public record Mapping(DeferredRegister<?> register) {
+        public void remap(String id, String newId) {
+            remap(ResourceLocation.parse(id), ResourceLocation.parse(newId));
         }
+
+        public void remap(ResourceLocation id, ResourceLocation object) {
+            register.addAlias(id, object);
+        }
+    }
+
+    private static void fixItems(@NotNull Mapping mapping) {
+        mapping.remap("vampiresdelight:eye_toast", "vampiresdelight:eye_croissant");
+        mapping.remap("vampiresdelight:wine_glass", "vampiresdelight:blood_wine_glass");
     }
 }