+++ src/main/java/net/grid/vampiresdelight/common/item/DaisyTeaItem.java
@@ -14,7 +14,6 @@ import net.minecraft.world.level.Level;
 import vectorwing.farmersdelight.common.Configuration;
 import vectorwing.farmersdelight.common.item.DrinkableItem;
 
-import javax.annotation.Nullable;
 import java.util.List;
 
 public class DaisyTeaItem extends DrinkableItem {
@@ -28,7 +27,7 @@ public class DaisyTeaItem extends DrinkableItem {
     }
 
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable Level level, List<Component> tooltip, TooltipFlag isAdvanced) {
+    public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltip, TooltipFlag isAdvanced) {
         Player player = VampirismMod.proxy.getClientPlayer();
 
         if (Configuration.FOOD_EFFECT_TOOLTIP.get() && player != null && VDEntityUtils.canConsumeHumanFood(player)) {