+++ src/main/java/net/grid/vampiresdelight/common/item/HardtackItem.java
@@ -1,5 +1,6 @@
 package net.grid.vampiresdelight.common.item;
 
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.food.FoodProperties;
 import net.minecraft.world.item.ItemStack;
 
@@ -9,7 +10,7 @@ public class HardtackItem extends HunterConsumableItem {
     }
 
     @Override
-    public int getUseDuration(ItemStack pStack) {
+    public int getUseDuration(ItemStack stack, LivingEntity entity) {
         return 48;
     }
 }