+++ src/main/java/net/grid/vampiresdelight/common/loot/modifier/VDAddLootTableModifier.java
@@ -1,37 +1,36 @@
 package net.grid.vampiresdelight.common.loot.modifier;
 
 import com.google.common.base.Suppliers;
-import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
 import net.grid.vampiresdelight.common.VDConfiguration;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraftforge.common.loot.IGlobalLootModifier;
-import net.minecraftforge.common.loot.LootModifier;
+import net.neoforged.neoforge.common.loot.AddTableLootModifier;
 
 import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.function.Supplier;
 
 import static net.minecraft.world.level.storage.loot.LootTable.createStackSplitter;
 
-/**
- * Credits to Commoble for this implementation!
- * VDAddLootTableModifier is Farmer's Delight's AddLootTableModifier adapted for Vampire's Delight
- */
-public class VDAddLootTableModifier extends LootModifier {
-    public static final Supplier<Codec<VDAddLootTableModifier>> CODEC = Suppliers.memoize(() ->
-            RecordCodecBuilder.create(inst -> codecStart(inst)
-                    .and(ResourceLocation.CODEC.fieldOf("lootTable").forGetter((m) -> m.lootTable))
+@SuppressWarnings("deprecation")
+@ParametersAreNonnullByDefault
+public class VDAddLootTableModifier extends AddTableLootModifier {
+    public static final Supplier<MapCodec<VDAddLootTableModifier>> CODEC = Suppliers.memoize(() ->
+            RecordCodecBuilder.mapCodec(inst -> codecStart(inst)
+                    .and(ResourceKey.codec(Registries.LOOT_TABLE).fieldOf("lootTable").forGetter((m) -> m.lootTable))
                     .apply(inst, VDAddLootTableModifier::new)));
 
-    private final ResourceLocation lootTable;
+    private final ResourceKey<LootTable> lootTable;
 
-    protected VDAddLootTableModifier(LootItemCondition[] conditionsIn, ResourceLocation lootTable) {
-        super(conditionsIn);
+    protected VDAddLootTableModifier(LootItemCondition[] conditionsIn, ResourceKey<LootTable> lootTable) {
+        super(conditionsIn, lootTable);
         this.lootTable = lootTable;
     }
 
@@ -39,14 +38,8 @@ public class VDAddLootTableModifier extends LootModifier {
     @Override
     protected ObjectArrayList<ItemStack> doApply(ObjectArrayList<ItemStack> generatedLoot, LootContext context) {
         if (VDConfiguration.GENERATE_VD_CHEST_LOOT.get()) {
-            LootTable extraTable = context.getResolver().getLootTable(this.lootTable);
-            extraTable.getRandomItemsRaw(context, createStackSplitter(context.getLevel(), generatedLoot::add));
+            context.getResolver().get(Registries.LOOT_TABLE, this.lootTable).ifPresent((extraTable) -> extraTable.value().getRandomItemsRaw(context, createStackSplitter(context.getLevel(), generatedLoot::add)));
         }
         return generatedLoot;
-    }
-
-    @Override
-    public Codec<? extends IGlobalLootModifier> codec() {
-        return CODEC.get();
     }
 }