+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.8
+
+### Additions
+- Ported to **1.21 NeoForge**;
+- Tooltips modified and added:
+  - **Spirit Lantern**'s tooltip is now hidden in order to save space on screen. Hold shift to view it;
+  - **Dandelion Beer Bottle** and **Blood Wine Bottle** now have a tooltip explaining how to pour drinks. Hold shift to view it;
+
 ## 0.1.7
 
 ### Additions